requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
python-dotenv==1.0.1
reportlab==4.2.2
pymysql==1.1.1